Add patterned UDL load combinations
